Move helper functions for test to `helpers_test.go`
