+++ copyandpaste_test.go
@@ -1,68 +1,8 @@
 package csvi_test
 
 import (
-	"errors"
-	"flag"
-	"fmt"
-	"io"
-	"os"
-	"path/filepath"
-	"strings"
 	"testing"
-
-	"github.com/mattn/go-colorable"
-
-	"github.com/hymkor/csvi/csviapp"
 )
-
-func newTestOptions(args ...string) (*csviapp.Options, error) {
-	flagSet := flag.NewFlagSet("test", flag.ContinueOnError)
-	opt := csviapp.NewOptions().Bind(flagSet)
-	return opt, flagSet.Parse(args)
-}
-
-func testRun(t *testing.T, dataSource io.Reader, args ...string) {
-	t.Helper()
-	instance, err := newTestOptions(args...)
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-	var ttyOut io.Writer = io.Discard
-	if testing.Verbose() {
-		disable := colorable.EnableColorsStdout(nil)
-		if disable != nil {
-			defer disable()
-		}
-		ttyOut = colorable.NewColorableStderr()
-	}
-	err = instance.RunInOut(dataSource, ttyOut)
-	if err != nil && !errors.Is(err, io.EOF) {
-		t.Fatal(err.Error())
-	}
-}
-
-func checkResult(t *testing.T, path, expect string) {
-	t.Helper()
-	bin, err := os.ReadFile(path)
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-	result := string(bin)
-	if expect != result {
-		t.Fatalf("Expect %#v, but %#v", expect, result)
-	}
-}
-
-func testCase(t *testing.T, source, process, result string, options ...string) {
-	t.Helper()
-	path := filepath.Join(t.TempDir(), "test.csv")
-	args := make([]string, 0, len(options)+3)
-	args = append(args, options...)
-	args = append(args, "-auto")
-	args = append(args, fmt.Sprintf("%s|w|%s|q|y", process, path))
-	testRun(t, strings.NewReader(source), args...)
-	checkResult(t, path, result)
-}
 
 func TestDeleteCell(t *testing.T) { // `x`
 	src := "あ,い,う,え,お\nか,き,く,け,こ"
